allow change uodate rest

+++ biodataware/api/groups/views.py
@@ -39,6 +39,11 @@ class GroupDetail(APIView):
 
     def delete(self, request, pk, format=None):
         group = get_object_or_404(Group, pk=pk)
+        if group.groupresearcher_set:
+            return Response({'detail': 'group not added! The group contains researcher(s).'}, status=status.HTTP_400_BAD_REQUEST)
+        if group.groupcontainer_set:
+            return Response({'detail': 'group not added! The group contains container(s).'},
+                            status=status.HTTP_400_BAD_REQUEST)
         group.delete()
         return Response({'detail': 'group deleted!'})
